Implement tests for * 🗹 long * 🗹 short * 🗹 Single (float) * =------>>> 🗹 Double * ☐ Decimal * ☐ Guid * ☐ DateTime * ☐ AnyOtherType
🗹

+++ src/Vogen/Templates/Decimal/Decimal_DapperTypeHandler.cs
@@ -12,7 +12,7 @@
                 {
                     int intValue => new VOTYPE(intValue),
                     long longValue when longValue < int.MaxValue => new VOTYPE((int)longValue),
-                    string stringValue when !string.IsNullOrEmpty(stringValue) && int.TryParse(stringValue, out var result) => new VOTYPE(result),
+                    string stringValue when !string.IsNullOrEmpty(stringValue) && decimal.TryParse(stringValue, out var result) => new VOTYPE(result),
                     _ => throw new System.InvalidCastException($"Unable to cast object of type {value.GetType()} to VOTYPE"),
                 };
             }
